+++ source/isaaclab/isaaclab/sensors/ray_caster/ray_caster.py
@@ -13,7 +13,7 @@ import numpy as np
 import torch
 import warp as wp
 
-import omni.physics.tensors.impl.api as physx
+import omni.physics.tensors.api as physx
 from pxr import Gf, Usd, UsdGeom, UsdPhysics
 
 import isaaclab.sim as sim_utils